Added Real SearchBar Component
This is a search bar that toggles between two versions of it using a bool value that you set in the constructor.

false -> if there's no query the list isn't displayed at the beginning.
true -> a list of all the items present in  the list will show at the beginning.

unsure how to make it accept different fields yet.